Refactor project to use SQLAlchemy instead of SQLite
The Python micro-web framework Flask was updated to use SQLAlchemy, a SQL toolkit and Object-Relational Mapping (ORM), instead of SQLite. This change includes updates to the Flask configuration, the addition of SQLAlchemy to project requirements, and other necessary file modifications to ensure compatibility. Also, some unused files were deleted, which aren't relevant with SQLAlchemy usage. This refactoring provides a more robust way to interact with databases, with added support for transaction and database session management.